+++ Services/Model/User.cs
@@ -5,17 +5,8 @@ namespace UI.Services.Model
 {
     public class User
     {
-        public User() { }
-
-        public User(string username, string password)
-        {
-            Username = username;
-            Password = password;
-        }
-
         public string Username { get; set; }
-        public string Password { get; set; }
-
+        public string Token { get; set; }
 
         public override string ToString() { return $"{{ Username: {Username} }}"; }
     }
